Shorten Commit Message
The subject line of a git commit message is 50 chars, the word "references" is 10, which means the current code uses 20% of the subject line per issue referenced before you can even get to meat of your subject line. Shortening to "re" allows the user to tag multiple issues in a commit and still squeeze in a meaningful subject.